refactor: extract shared stripHtml and escapeCSVValue into stringUtils
Removes duplicate private methods from AssessmentReportHelper and
ResultsBuilder by moving them to a shared stringUtils module.
